+++ wxc_sdk/as_api.py
@@ -1306,7 +1306,7 @@ class AsEventsApi(AsApiChild, base='events'):
         if resource is not None:
             params['resource'] = enum_str(resource)
         if type_ is not None:
-            params['type'] = enum_str(type)
+            params['type'] = enum_str(type_)
         if actor_id is not None:
             params['actorId'] = actor_id
         if has_attachments is not None:
@@ -1351,7 +1351,7 @@ class AsEventsApi(AsApiChild, base='events'):
         if resource is not None:
             params['resource'] = enum_str(resource)
         if type_ is not None:
-            params['type'] = enum_str(type)
+            params['type'] = enum_str(type_)
         if actor_id is not None:
             params['actorId'] = actor_id
         if has_attachments is not None: